requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
PyPDF2
sqlalchemy
pg8000
python-jose[cryptography]
passlib
bcrypt
httpx